Moved the transform unwrapping from VSAbstractFilter to SIMPLVtkBridge when dealing with vtkImageData.  Other geometries are left without wrapping their transforms.
VSFilterViewSettings use actor position, rotation, and scale when dealing with 2D image data and outline representation.  3D data uses the transformed output port when rendering.

The vtkOutlineFilter was moved to VSFilterViewSettings from VSAbstractFilter.  Attempts were made to transform the data through VSAbstractFilter before creating the outline, but there were problems with this approach and it was scrapped.

Cleaned up VSController when importing data through a VSDataSetFilter.  VSDataSetFilter was, for some reason, not having a parent and cannot have children.